разбил файл main.js на layout.js и main.js теперь каждая страница будет иметь layout.js, для корректной работы шапки и футера остальные скрипты будут подключаться отдельно и передаваться с блоком cshtml
