+++ MyTotto.Web/Views/Home/Index.cshtml
@@ -8,3 +8,5 @@
 @await Html.PartialAsync("_Products", Model.Products)
 @await Html.PartialAsync("_Discounts", Model.Promos)
 @await Html.PartialAsync("_NewProducts", Model.PromoProducts)
+
+<script src="~/js/main.js" asp-append-version="true"></script>
