+++ MyTotto.Web/Views/Shared/_Layout.cshtml
@@ -32,7 +32,7 @@
 
     @await Html.PartialAsync("_Footer")
 
-    <script src="~/js/main.js" asp-append-version="true"></script>
+    <script src="~/js/layout.js" asp-append-version="true"></script>
 
     @await RenderSectionAsync("Scripts", required: false)
 </body>
